add rule export feature
