+++ .gitignore
@@ -11,3 +11,4 @@ __pycache__/
 dist/
 sophos_firewall_audit/results_html*
 firewalls.yaml
+/rule_export*
