+++ sophos_firewall_audit/postaudit_web.py
@@ -102,7 +102,7 @@ if __name__ == "__main__":
     
     current_directory = os.path.dirname(os.path.abspath(__file__))
     parent_directory = os.path.dirname(current_directory)
-    with open(os.path.join(parent_directory, "results.json"), "r") as fn:
+    with open(os.path.join(parent_directory, "results.json"), "r", encoding="utf-8") as fn:
         results = json.loads(fn.read())
 
     html_table = parse_results(results)
