+++ sophos_firewall_audit/sophosfirewallaudit.py
@@ -27,78 +27,21 @@ NAUTOBOT_TOKEN = Nautobot API Token
 """
 # Patching prettytable to disable escaping
 # https://github.com/jazzband/prettytable/issues/40
-import html
-html.escape = lambda *args, **kwargs: args[0]
-from prettytable import PrettyTable
-from prettytable import ALL
+import logging
+from datetime import datetime
+import argparse
+import os
+import sys
+import yaml
 from pynautobot import api
 from sophosfirewall_python.firewallapi import SophosFirewall
 from sophos_firewall_audit.auth import get_credential
-from sophos_firewall_audit import rules
-import os
-import logging
-from datetime import datetime
-import json
-import argparse
-import yaml
-import pkg_resources
-from jinja2 import Environment, PackageLoader, Template, select_autoescape
-from rich.logging import RichHandler
-from rich.highlighter import RegexHighlighter
-from rich.theme import Theme
-from rich.console import Console
+from sophos_firewall_audit.logging_config import LoggingSetup
+from sophos_firewall_audit.audit import run_audit, generate_audit_output
+from sophos_firewall_audit.rule_export import export_rules, generate_rule_output
 
-
-
-
-class DeviceNameHighlighter(RegexHighlighter):
-    """Apply style to the device name."""
-
-    base_style = "style."
-    highlights = [r"(?P<hostname>[a-z]+-[a-z]+-[a-z]+-[a-z]+-\S+)", r"(?P<hostname>[a-z]+-[a-z]+-[a-z]+-[a-z]+-[a-z]+-\S+)"]
-
-theme = Theme({"style.hostname": "magenta"})
-console = Console(theme=theme)
-FORMAT = '%(message)s'
-logging.basicConfig(level=logging.INFO, format=FORMAT, handlers=[RichHandler(console=console, 
-                                                                             highlighter=DeviceNameHighlighter(),
-                                                                             show_path=False,
-                                                                             omit_repeated_times=False)])
-
-
-def update_status_dict(result, status_dict, firewall_name):
-    """Update overall status counters in the status_dict
-
-    Args:
-        result (dict): Rule evaluation result
-        status_dict (dict): Pass/Fail status tracker
-        firewall_name (str): Firewall hostname
-
-    Returns:
-        dict: Returns the passed in status_dict with the counters updated
-    """
-    status_dict[firewall_name]["success_ct"] += result["pass_ct"]
-    status_dict[firewall_name]["failed_ct"] += result["fail_ct"]
-    logging.info(f"{firewall_name}: Success: {status_dict[firewall_name]['success_ct']} Failed: {status_dict[firewall_name]['failed_ct']}")
-    return status_dict
-
-def process_rule(method, settings, log_msg, fw_obj, status_dict):
-    """Process an evaluation rule. 
-
-    Args:
-        method (function): The rule function to be executed
-        settings (dict): The expected result from the settings YAML
-        log_msg (str): Log message to be printed
-        fw_obj (obj): SophosFirewall object
-        status_dict (dict): Pass/Fail status counters
-
-    Returns:
-        dict: Task output and status_dict with updated counters
-    """
-    logging.info(f"{fw_obj.hostname}: {log_msg}")
-    result = method(fw_obj, fw_obj.hostname, settings)
-    status_dict = update_status_dict(result, status_dict, fw_obj.hostname)
-    return {"result": result, "output": result["output"], "status_dict": status_dict}
+logging_setup = LoggingSetup()
+logger = logging.getLogger(__name__)
 
 def nb_graphql_query(query, verify):
     """Query Nautobot using GraphQL returning a list of device names
@@ -155,6 +98,7 @@ def all_devices_query(environ):
     templ = environ.get_template("all_devices_query.j2")
     return templ.render()
 
+
 def create_dir(dirpath):
     try:
         os.mkdir(dirpath)
@@ -163,10 +107,6 @@ def create_dir(dirpath):
     return
 
 def main():
-    env = Environment(
-        loader=PackageLoader("sophos_firewall_audit"),
-        autoescape=select_autoescape()
-    )
 
     parser = argparse.ArgumentParser()
     group1 = parser.add_mutually_exclusive_group(required=True)
@@ -176,8 +116,14 @@ def main():
     parser.add_argument("-d","--disable_verify", help="Disable certificate checking for Nautobot", action="store_false", default=True)
     parser.add_argument("-s", "--settings_file", help="Audit settings YAML file", default="audit_settings.yaml")
     parser.add_argument("-u", "--use_vault", help="Use HashiCorp Vault to retrieve credentials", action="store_true", default=False)
+    parser.add_argument("-r", "--rule_export", help="Export rules for offline viewing", action="store_true", default=False)
 
     args = parser.parse_args()
+
+    if args.use_nautobot and not args.query_file:
+        parser.error('--query_file is required when --use_nautobot is specified.')
+
+    
     logging.info("Starting Sophos Firewall audit")
 
     if args.use_nautobot:
@@ -191,6 +137,11 @@ def main():
 
     if args.use_vault:
         logging.info ("Retrieving credentials from Vault...")
+        try:
+            os.environ["VAULT_ADDR"]
+        except KeyError:
+            logging.error("Missing VAULT_ADDR environment variable!")
+            sys.exit(1)
         fw_password = get_credential(
             mount_point=os.environ['VAULT_MOUNT_POINT'],
             secret_path=os.environ['VAULT_SECRET_PATH'],
@@ -202,18 +153,24 @@ def main():
         fw_username = os.environ.get("FW_USERNAME")
         fw_password = os.environ.get("FW_PASSWORD")
         
-  
     status_dict = {}
 
-
-    create_dir("results_html_local")
-    create_dir("results_html_web")
-
     dt = datetime.now()
-    local_dirname = os.path.join("results_html_local", f"audit-results-{dt.strftime('%Y-%m-%d-%H%M%S')}")
-    web_dirname = os.path.join("results_html_web", f"audit-results-{dt.strftime('%Y-%m-%d-%H%M%S')}")
-    os.mkdir(local_dirname)
-    os.mkdir(web_dirname)
+    if not args.rule_export:
+        create_dir("results_html_local")
+        create_dir("results_html_web")      
+        local_dirname = os.path.join("results_html_local", f"audit-results-{dt.strftime('%Y-%m-%d-%H%M%S')}")
+        web_dirname = os.path.join("results_html_web", f"audit-results-{dt.strftime('%Y-%m-%d-%H%M%S')}")
+        os.mkdir(local_dirname)
+        os.mkdir(web_dirname)
+    elif args.rule_export:
+        logging.info("Rule export flag detected, initiating rule export (audit will be skipped!)")
+        create_dir("rule_export_local")
+        create_dir("rule_export_web")
+        local_dirname = os.path.join("rule_export_local", f"rule-export-{dt.strftime('%Y-%m-%d-%H%M%S')}")
+        web_dirname = os.path.join("rule_export_web", f"rule-export-{dt.strftime('%Y-%m-%d-%H%M%S')}")
+        os.mkdir(local_dirname)
+        os.mkdir(web_dirname)
 
     for firewall in firewalls:
         fw = SophosFirewall(
@@ -223,206 +180,23 @@ def main():
             port=firewall['port'],
             verify=False
         )
-
         try:
             fw.login()
         except Exception as Error:
             logging.error(f"Error connecting to firewall {firewall['hostname']}: {Error}")
             continue
-
-        with open(args.settings_file, "r") as fn:
-            templ = Template(source=fn.read())
-            rendered = templ.render({"firewall_hostname": fw.hostname.split(".")[0]})
-            audit_settings = yaml.safe_load(rendered)
-
-        results = []
-        output = []
-        firewall_name = firewall["hostname"]
-        status_dict[firewall_name] = {
-            "success_ct": 0,
-            "failed_ct": 0
-        }
-
-        logging.info(f"{firewall_name}: Begin Audit")
-
-        rule_list = [
-            {
-                "method": rules.eval_access_list, 
-                "settings": audit_settings["access_acl"],
-                "log_msg": "Evaluate Access ACL"
-            },
-            {
-                "method": rules.eval_central_mgmt, 
-                "settings": audit_settings,
-                "log_msg": "Evaluate Central Management"
-            },
-            {
-                "method": rules.eval_device_access_profile,
-                "settings": audit_settings["device_access_profile"],
-                "log_msg": "Evaluate Device Access Profiles"
-            },
-            {
-                "method": rules.eval_admin_services,
-                "settings": audit_settings["admin_services"],
-                "log_msg": "Evaluate WAN Zone Admin Services"
-            },
-            {
-                "method": rules.eval_admin_authen,
-                "settings": audit_settings["authen_servers"],
-                "log_msg": "Evaluate Authentication Servers"
-            },
-            {
-                "method": rules.eval_malware_protection,
-                "settings": audit_settings["malware_protection"],
-                "log_msg": "Evaluate Malware Protection Antivirus Engine"
-            },
-            {
-                "method": rules.eval_atp,
-                "settings": audit_settings["threat_protection"],
-                "log_msg": "Evaluate Advanced Threat Protection (ATP)"
-            },
-            {
-                "method": rules.eval_ips_policies,
-                "settings": audit_settings["ips_policies"],
-                "log_msg": "Evaluate IPS Policies"
-            },
-            {
-                "method": rules.eval_hostgroups,
-                "settings": audit_settings["host_groups"],
-                "log_msg": "Evaluate Host Groups"
-            },
-            {
-                "method": rules.eval_syslog,
-                "settings": audit_settings["syslog"],
-                "log_msg": "Evaluate Syslog Settings"
-            },
-            {
-                "method": rules.eval_notifications,
-                "settings": audit_settings,
-                "log_msg": "Evaluate Notifications Settings"
-            },
-            {
-                "method": rules.eval_notification_list,
-                "settings": audit_settings,
-                "log_msg": "Evaluate Notification List Settings"
-            },
-            {
-                "method": rules.eval_backup,
-                "settings": audit_settings,
-                "log_msg": "Evaluate Scheduled Backup Settings"
-            },
-            {
-                "method": rules.eval_certificate,
-                "settings": audit_settings,
-                "log_msg": "Evaluate Certificate Settings"
-            },
-            {
-                "method": rules.eval_loginsecurity,
-                "settings": audit_settings,
-                "log_msg": "Evaluate Login Security"
-            },
-            {
-                "method": rules.eval_dns_servers,
-                "settings": audit_settings,
-                "log_msg": "Evaluate DNS Servers"
-            },
-            {
-                "method": rules.eval_smtp_protection,
-                "settings": audit_settings,
-                "log_msg": "Evaluate SMTP Protection"
-            },
-            {
-                "method": rules.eval_snmpv3,
-                "settings": audit_settings,
-                "log_msg": "Evaluate SNMPv3"
-            },
-            {
-                "method": rules.eval_time,
-                "settings": audit_settings['time'],
-                "log_msg": "Evaluate Time Settings"
-            }
-        ]
-        for rule in rule_list:
-            result = process_rule(rule["method"], rule["settings"], rule["log_msg"], fw, status_dict)
-            output += result["output"]
-            status_dict = result["status_dict"]
-            results.append(result)
         
-        table = PrettyTable()
-        table.hrules = ALL
-        table.field_names = ["Test Name", "UI Location", "Object", "Expected", "Actual", "Result"]
-        table.add_rows(output)
-        table.valign = "m"
-        table.align["Expected"] = "l"
-        table.align["Actual"] = "l"
-
-        template = env.get_template("results.j2")
-        result_html = template.render(firewall_name=firewall_name, table=table.get_html_string(format=True, escape_data=False))
-
-        for dirname in [local_dirname, web_dirname]:
-            with open (f"{os.path.join(dirname, firewall_name)}.html", "w", encoding="utf-8") as fn:
-                fn.write(result_html)
-
-            with open(f"{os.path.join(dirname, firewall['hostname'])}.json", "w", encoding="utf-8") as fn:
-                fn.write(json.dumps(results, indent=4))
-
-    template = env.get_template("index.j2")
-
-    firewall_list = [firewall['hostname'] for firewall in firewalls]
-
-    with open("results.json", "w", encoding="utf-8") as fn:
-        fn.write(json.dumps(status_dict))
-
-    for dirname in [local_dirname, web_dirname]:
-        index_html = template.render(status_dict=status_dict, dirname=dirname)
-
-        with open(os.path.join(dirname, 'index.html'), "w", encoding="utf-8") as fn:
-            fn.write(index_html)
-
-        template_file = pkg_resources.resource_filename("sophos_firewall_audit", os.path.join("templates", "home_index.j2"))
-
-        if "web" in dirname:
-            try:
-                with open(os.path.join("results_html_web", "index.html"), "r", encoding="utf-8") as fn:
-                    home_html = fn.readlines()
-            except FileNotFoundError:
-                with open(template_file, "r") as fn:
-                    home_html = fn.readlines()
-                with open(os.path.join("results_html_web", "index.html"), "w", encoding="utf-8") as fn:
-                    fn.writelines(home_html)
-        else:
-            try:
-                with open(os.path.join("results_html_local", "index.html"), "r", encoding="utf-8") as fn:
-                    home_html = fn.readlines()
-            except FileNotFoundError:
-                with open(template_file, "r") as fn:
-                    home_html = fn.readlines()
-                with open(os.path.join("results_html_local", "index.html"), "w", encoding="utf-8") as fn:
-                    fn.writelines(home_html)                  
-
-        updated_home_html = []
-
-        for line in home_html:
-            if "<h1>Firewall Audit</h1>" in line:
-                updated_home_html.append(line)
-                if "web" in dirname:
-                    updated_home_html.append(
-                        f'<a style="text-align: left;" href="/{dirname.split(os.sep)[1]}/index.html">{dirname.split(os.sep)[1]}</a><br/>\n'
-                    )
-                    
-                else:
-                    updated_home_html.append(
-                        f'<a style="text-align: left;" href="file:{dirname.split(os.sep)[1]}/index.html">{dirname.split(os.sep)[1]}</a><br/>\n'
-                    )
-            else:
-                updated_home_html.append(line)
-
-        if "web" in dirname:
-            with open(os.path.join("results_html_web", "index.html"), "w", encoding="utf-8") as fn:
-                fn.writelines(updated_home_html)
-        else:
-            with open(os.path.join("results_html_local", "index.html"), "w", encoding="utf-8") as fn:
-                fn.writelines(updated_home_html)
+        if not args.rule_export:
+            status_dict = run_audit(args, fw, firewall, status_dict, local_dirname, web_dirname)
+        elif args.rule_export:
+            export_rules(fw, firewall, local_dirname, web_dirname)
+            logging.info(f"{firewall['hostname']}: Rule export completed successfully!")
+    
+    if not args.rule_export:
+        generate_audit_output(status_dict, local_dirname, web_dirname)
+    elif args.rule_export:
+        generate_rule_output(firewalls, local_dirname, web_dirname)
+    
 
 if __name__ == "__main__":
     main()
