+++ sophos_firewall_audit/templates/audit_index.j2
@@ -35,7 +35,7 @@ th, td {
 }
 </style>
 <div>
-    <h1>Firewall Audit Report</h1>
+    <h1>{{ title }}</h1>
     <table frame="box" rules="all">
         <tbody>
             <tr>
